# Feat/optimized sync watcher