--- a/packages/hop-node/src/constants/constants.ts
+++ b/packages/hop-node/src/constants/constants.ts
@@ -160,3 +160,19 @@ export const FinalityTagForChain: Record<string, string> = {
   [Chain.Nova]: FinalityTag.Safe,
   [Chain.PolygonZk]: FinalityTag.Safe
 }
+
+export const SyncIterationMultiplier: Record<string, number> = {
+  [Chain.Ethereum]: 1,
+  [Chain.Gnosis]: 2,
+  [Chain.Polygon]: 2,
+  [Chain.Optimism]: 1,
+  [Chain.Arbitrum]: 1,
+  [Chain.Base]: 1,
+  [Chain.Nova]: 2,
+  [Chain.PolygonZk]: 1
+}
+
+export const ShouldRelayL1ToL2Message: Record<string, boolean> = {
+  [Chain.Arbitrum]: true,
+  [Chain.Nova]: true
+}