Optimizes the sync watcher with a few changes and new features.

* Default sync cycle time updated to 30 sec (from 60 sec)
* `TransferSent` (and `TransferSentToL2` on relayable chains) events are retrieved every cycle while all others are retrieved on a different cadence (currently 60x the default cycle time, or 30 mins).
* Introduce a multiplier `SyncIterationMultiplier` for less-used chains. This multiplier updates the default cycle time (30 sec) to something like 60 sec.
* Minor optimization: the bonder was making a `getLogs` call every cycle, even if the latest known block was already at the head. Now, if the chain head is already known by the bonder, they will not make a `getLogs` call anymore.